_KERNEL_MODE  delete/changed

+++ src/AKNet.MSQuic/msquic1/platform/platform_winuser.cs
@@ -100,7 +100,7 @@ namespace MSQuic1
         //    }
         //}
 
-        static void SetThreadAffinity(ushort nProcessorId)
+        static void SetThreadAffinity(int nProcessorId)
         {
             Thread.BeginThreadAffinity();
